requirements.txt
Django==5.2.7
djangorestframework==3.15.2
django-filter==24.2
PyMySQL==1.1.1
python-dotenv==1.0.0
pandas==2.2.3
openpyxl==3.1.5
